splits off components to own files, adds back button
